mc_api::mc_state_choose_request() updated
